requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.6
distro==1.9.0
exceptiongroup==1.2.1
faiss-cpu==1.8.0
faiss-gpu==1.7.2
fastapi==0.110.3
Flask==3.0.3
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.2.2
langchain-cli==0.0.24
langchain-community==0.2.3
langchain-core==0.2.4
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langserve==0.2.1
langsmith==0.1.71
libcst==1.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
ngrok==1.3.0
numpy==1.26.4
openai==1.31.0
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyproject-toml==0.0.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.30
sse-starlette==1.8.2
starlette==0.37.2
tenacity==8.3.0
tiktoken==0.7.0
toml==0.10.2
tomlkit==0.12.5
tqdm==4.66.4
twilio==9.1.0
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.23.2
Werkzeug==3.0.3
yarl==1.9.4
